@@ -1,6 +1,6 @@
 {
   "name": "react-smtc-ui-utils",
-  "version": "1.4.5",
+  "version": "1.4.6",
   "description": "react-smtc-ui-utils React component",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -12,35 +12,36 @@
     "umd"
   ],
   "scripts": {
-    "build": "nwb build-react-component",
+    "build": "NODE_OPTIONS=--openssl-legacy-provider nwb build-react-component",
     "clean": "nwb clean-module && nwb clean-demo",
-    "start": "nwb serve-react-demo",
+    "start": "NODE_OPTIONS=--openssl-legacy-provider nwb serve-react-demo",
     "test": "nwb test-react",
     "test:coverage": "nwb test-react --coverage",
     "test:watch": "nwb test-react --server",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d demo/dist",
+    "deploy": "NODE_OPTIONS=--openssl-legacy-provider gh-pages -d demo/dist -r https://<token>@github.com/fireHedgehog/react-smtc-ui-utils.git",
     "docgen": "react-docgen src --pretty",
     "docgen-resolver": "react-docgen src --pretty --resolver findAllComponentDefinitions -o src/doc/doc.json"
   },
   "dependencies": {
     "lodash": "^4.17.21",
-    "prop-types": "^15.7.2",
-    "react-resize-detector": "^6.7.6"
+    "prop-types": "^15.8.1",
+    "react-resize-detector": "^10.0.1"
   },
   "devDependencies": {
-    "npm-check-updates": "^11.8.5",
-    "gh-pages": "^3.2.3",
-    "highlight.js": "^11.2.0",
-    "html-entities": "^2.3.2",
+    "npm-check-updates": "^16.14.18",
+    "gh-pages": "^6.1.1",
+    "highlight.js": "^11.9.0",
+    "html-entities": "^2.5.2",
     "nwb": "^0.25.2",
-    "react": "^17.0.2",
-    "react-docgen": "^5.4.0",
-    "react-dom": "^17.0.2",
-    "react-highlight": "^0.14.0",
-    "react-router-dom": "^5.3.0",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.0.4"
+    "react": "^18.2.0",
+    "react-docgen": "^7.0.3",
+    "react-dom": "^18.2.0",
+    "react-highlight": "^0.15.0",
+    "react-router-dom": "^6.22.3",
+    "semantic-ui-css": "^2.5.0",
+    "semantic-ui-react": "^2.1.5",
+    "yarn-upgrade-all": "^0.7.2"
   },
   "author": "rockychen",
   "license": "MIT",
