@@ -1,10 +1,7 @@
-import React, {Component} from 'react'
+import React from 'react'
 import {render} from 'react-dom'
 import 'semantic-ui-css/semantic.min.css';
-import {
-    HashRouter as Router,
-    Route,
-} from "react-router-dom";
+import {HashRouter as Router, Route, Routes} from "react-router-dom";
 import MainMenu from "./menu/MainMenu";
 import {BasicTableDemo} from "./table/BasicTableDemo";
 import {Container} from "semantic-ui-react";
@@ -17,60 +14,27 @@ import "highlight.js/styles/idea.css";
 import PropsOfTable from "./table/PropsOfTable";
 import {TableResponsiveContainer} from "./table/TableResponsiveContainer";
 
-class Index extends Component {
 
-
-    render() {
-        return (
-            <Router>
-
-                <div>
-
-                    <MainMenu {...this.props}/>
-
-                    <Container>
-                        <MainTitle/>
-                        <Route exact path="/" component={HomePage}/>
-                        <Route path={`/react-smtc-ui-utils/start`} component={GetStart}/>
-                        <Route path={`/react-smtc-ui-utils/table`} component={BasicTableDemo}/>
-                        <Route path={`/react-smtc-ui-utils/Props`} component={PropsOfTable}/>
-                        <Route path={`/react-smtc-ui-utils/tableProps`} component={TablePropsContainer}/>
-                        <Route path={`/react-smtc-ui-utils/responsive`} component={TableResponsiveContainer}/>
-                        <Route path={`/react-smtc-ui-utils/others`} component={StepDemo}/>
-                    </Container>
-
-                </div>
-
-            </Router>
-        )
-    }
-}
-
-/*
-class Index extends Component {
-
-    render() {
-
-        return (
+const Index = () => {
+    return (
+        <Router>
             <div>
-                <Router>
-                    <div>
-                        <Route exact
-                               path={"/"}
-                               render={(props) => <Home {...props}  />}
-
-                        />
-                        <Route path={"/table"}
-                               render={(props) => <BasicTableDemo {...props}  />}
-
-                        />
-                    </div>
-                </Router>
+                <MainMenu/>
+                <Container>
+                    <MainTitle/>
+                    <Routes>
+                        <Route exact path="/" element={<HomePage/>}/>
+                        <Route path="/react-smtc-ui-utils/start" element={<GetStart/>}/>
+                        <Route path="/react-smtc-ui-utils/table" element={<BasicTableDemo/>}/>
+                        <Route path="/react-smtc-ui-utils/Props" element={<PropsOfTable/>}/>
+                        <Route path="/react-smtc-ui-utils/tableProps" element={<TablePropsContainer/>}/>
+                        <Route path="/react-smtc-ui-utils/responsive" element={<TableResponsiveContainer/>}/>
+                        <Route path="/react-smtc-ui-utils/others" element={<StepDemo/>}/>
+                    </Routes>
+                </Container>
             </div>
-        )
-    }
-}
-*/
+        </Router>
+    );
+};
 
-
-render(<Index/>, document.querySelector('#demo'))
+render(<Index/>, document.querySelector('#demo'));
